+++ test/pulsar/components/menu_test.exs
@@ -445,5 +445,43 @@ defmodule Pulsar.Components.MenuTest do
       refute html =~ "data-method="
       refute html =~ "data-csrf="
     end
+
+    test "method= with navigate= raises" do
+      assigns = %{}
+
+      assert_raise ArgumentError, ~r/:method cannot be used with :navigate or :patch/, fn ->
+        rendered_to_string(~H"""
+        <Menu.menu id="m" label="Primary">
+          <Menu.menu_item navigate="/sign-out" method="delete">Sign out</Menu.menu_item>
+        </Menu.menu>
+        """)
+      end
+    end
+
+    test "method= with patch= raises" do
+      assigns = %{}
+
+      assert_raise ArgumentError, ~r/:method cannot be used with :navigate or :patch/, fn ->
+        rendered_to_string(~H"""
+        <Menu.menu id="m" label="Primary">
+          <Menu.menu_item patch="/sign-out" method="delete">Sign out</Menu.menu_item>
+        </Menu.menu>
+        """)
+      end
+    end
+
+    test "method= on an action item does not reach the button" do
+      assigns = %{}
+
+      html =
+        rendered_to_string(~H"""
+        <Menu.menu id="m" label="Primary">
+          <Menu.menu_item phx-click="sign_out" method="delete">Sign out</Menu.menu_item>
+        </Menu.menu>
+        """)
+
+      assert html =~ "<button"
+      refute html =~ ~s(method="delete")
+    end
   end
 end